Blocks dispatchosglobal.yuanshen.com temporarily
Query logs to block the domain at least once.
Once the script finds a log entry indicating it has blocked the domain.
Then reenable.

On blocking and unblocking the domain, verify that the domain name is
blocked/unblocked.
It takes time for filter to take effect, so it polls whether it has
taken effect.